==> 2. Base Model Data.Rmd <==
---
title: "Base Model Data"
author: "Elisabeth Nelson"
date: "2022-12-06"
output: pdf_document
---
## Load Packages
```{r}

library(stats)
library(ggplot2)
library(dplyr)
library(tidyverse)
library(TSDT)
library(lubridate)
library(sp)
library(sf)

```

## Create an Aggregated DF With Distinct District Names 
```{r}
# read in MDR_Dengue_mpa
MDR_Dengue_map_restored <- readRDS("MDR_Dengue_map.RDS")


# create data frame
base_model_2001_2006_agg <- fortify(MDR_Dengue_map_restored) 

base_model_2001_2006_agg_2 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "An Giang" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh An Giang"))

base_model_2001_2006_agg_3 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Bến Tre" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Ben Tre"))

base_model_2001_2006_agg_4 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Đồng Tháp" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Dong Thanp"))

base_model_2001_2006_agg_5 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Hậu Giang" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Hau Giang"))

base_model_2001_2006_agg_6 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Kiên Giang" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Kien Giang"))

base_model_2001_2006_agg_7 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Long An" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Long An"))
 
base_model_2001_2006_agg_8 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Sóc Trăng" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Soc Trang")) 
         
base_model_2001_2006_agg_9 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Tiền Giang" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Tien Giang"))

base_model_2001_2006_agg_10 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Trà Vinh" & VARNAME_2 == "Chau Thanh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Chau Thanh", "Chau Thanh Tra Vinh"))

base_model_2001_2006_agg_11 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "An Giang" & VARNAME_2 == "Phu Tan") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Phu Tan", "Phu Tan An Giang")) 
      
base_model_2001_2006_agg_12 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Cà Mau" & VARNAME_2 == "Phu Tan") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Phu Tan", "Phu Tan Ca Mau"))

base_model_2001_2006_agg_13 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Đồng Tháp" & VARNAME_2 == "Cao Lanh" & district == "cao lãnh") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Cao Lanh", "Cao Lanh District"))

base_model_2001_2006_agg_14 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Đồng Tháp" & VARNAME_2 == "Cao Lanh" & ENGTYPE_2 == "City") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Cao Lanh", "Cao Lanh City"))

base_model_2001_2006_agg_15 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Đồng Tháp" & VARNAME_2 == "Hong Ngu" & district == "hồng ngự") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Hong Ngu", "Hong Ngu District"))

base_model_2001_2006_agg_16 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Đồng Tháp" & VARNAME_2 == "Hong Ngu" & ENGTYPE_2 == "City") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Hong Ngu", "Hong Ngu City"))

base_model_2001_2006_agg_17 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Hậu Giang" & VARNAME_2 == "Long My" & district == "long mỹ") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Long My", "Long My District"))

base_model_2001_2006_agg_18 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Hậu Giang" & VARNAME_2 == "Long My" & ENGTYPE_2 == "Town") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Long My", "Long My Town"))

base_model_2001_2006_agg_19 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Tiền Giang" & VARNAME_2 == "Cai Lay" & district == "cai lậy") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Cai Lay", "Cai Lay District"))

base_model_2001_2006_agg_20 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Tiền Giang" & VARNAME_2 == "Cai Lay" & ENGTYPE_2 == "Town") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Cai Lay", "Cai Lay Town"))

base_model_2001_2006_agg_21 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Trà Vinh" & VARNAME_2 == "Duyen Hai" & district == "duyên hải") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Duyen Hai", "Duyen Hai District"))

base_model_2001_2006_agg_22 <- base_model_2001_2006_agg %>%
  filter(NAME_1 == "Trà Vinh" & VARNAME_2 == "Duyen Hai" & ENGTYPE_2 == "Town") %>%
  mutate(VARNAME_2 = str_replace(VARNAME_2, "Duyen Hai", "Duyen Hai Town"))

base_model_2001_2006_agg_filter <- rbind(base_model_2001_2006_agg, base_model_2001_2006_agg_2, base_model_2001_2006_agg_3, base_model_2001_2006_agg_4, base_model_2001_2006_agg_5, base_model_2001_2006_agg_6, base_model_2001_2006_agg_7, base_model_2001_2006_agg_8, base_model_2001_2006_agg_9, base_model_2001_2006_agg_10, base_model_2001_2006_agg_11, base_model_2001_2006_agg_12, base_model_2001_2006_agg_13, base_model_2001_2006_agg_14, base_model_2001_2006_agg_15, base_model_2001_2006_agg_16, base_model_2001_2006_agg_17, base_model_2001_2006_agg_18, base_model_2001_2006_agg_19, base_model_2001_2006_agg_20, base_model_2001_2006_agg_21, base_model_2001_2006_agg_22)

names_to_delete <- c("Chau Thanh", "Phu Tan", "Cao Lanh", "Hong Ngu", "Long My", "Cai Lay", "Duyen Hai")

base_model_2001_2006_agg_filter <- base_model_2001_2006_agg_filter %>%
  filter(!(VARNAME_2 %in% names_to_delete))

# rotate month variables into one column
base_model_2001_2006_filter_long <- base_model_2001_2006_agg_filter %>%
  pivot_longer(cols = starts_with("m"), 
               values_to = "m_DHF_cases", 
               names_to = "month") %>%
  mutate(month = as.numeric(gsub('m','', month)),
         m_DHF_cases = as.numeric(m_DHF_cases))

# ensure all districts are listed through all the years
agg1 <- base_model_2001_2006_filter_long %>%
  #bind_rows() %>% 
  #group_by(VARNAME_2, year, month) %>%
  #summarize(m_DHF_cases = n())  %>%
  #ungroup  %>%
  mutate(month=as.factor(month),
 year=as.factor(year),
 district =as.factor(VARNAME_2)) %>%
  tidyr::complete(year, month, VARNAME_2, fill=list(m_DHF_cases=NA)) #fills NA

# fill-in missing geometries
base_model_geometries <- base_model_2001_2006_filter_long %>%
  dplyr::select(VARNAME_2, geometry) 

base_model_geometries <- base_model_geometries %>%
  filter(VARNAME_2 == unique(VARNAME_2)) %>%
  distinct(VARNAME_2, .keep_all = T)

agg2 <- left_join(agg1, base_model_geometries, by = c("VARNAME_2" = "VARNAME_2"))

agg3 <- agg2 %>%
  select(!(geometry.x)) %>%
  rename("geometry" = geometry.y)

# upload MDR_2
MDR_2 <- st_read(dsn = '/Users/zoe/Documents/Yale EMD/Rotation 1/Dengue - Vietnam/dengue_exploration/Data/MDR Updated map/VMD_districts_2022.shp')
MDR_2 %>%
  mutate(MDR_2, VARNAME_2 = as.character(VARNAME_2)) %>%
  mutate(MDR_2, ENGTYPE_2 = as.character(ENGTYPE_2)) 

missing_long_my <- MDR_2$geometry[MDR_2$NAME_2 == "Long Mỹ (Thị xã)"]

agg3$geometry[agg3$VARNAME_2 == "Long My Town"] <- missing_long_my

# put in final form
base_model_final <- agg3 %>%
  mutate(month = as.numeric(month),
    m_DHF_cases = as.numeric(m_DHF_cases))

```

## Save as a .RDS
```{r}

saveRDS(base_model_final, "base_model_final.RDS")

```
