==> 4. Base_Model_Function_for_Cluster.Rmd <==
---
title: "Base Model Function for Cluster"
author: "Elisabeth Nelson"
date: "2022-12-11"
output: pdf_document
---
## Load Packages
```{r}

library(MCMCpack)
library(stats)
library(CARBayes) 
library(rgdal) 
library(RColorBrewer) 
library(ggplot2)
library(rgeos) 
library(maptools) 
library(spdep)  
library(ggmap)  
library(sf)
library(dplyr)
library(tidyverse)
library(CARBayesST)
library(TSDT)
library(lubridate)

```

## Load Needed Files
```{r}
base_model_final <- readRDS("base_model_final.RDS")
neighb.mat_2 <- readRDS("neighb.mat_2.RDS")

```

## Function for generating dataset
```{r}

#initially, fit through nov 2006, set dec 2006 to NA
#'2006-12-01'

base_model_final <- base_model_final %>%
  mutate(year = as_factor(year)) %>%
  filter(VARNAME_2 != "Kien Hai", 
         VARNAME_2 != "Phu Quoc") 


advance_missing_func <- function(N_advance){
      
      base_model_for_model_1 <- base_model_final %>% 
        mutate(year = unfactor(year),
               #monthN = as.numeric(gsub('m','', month)),
               date = as.Date(paste(year, month, '01', sep='-')),
               pop=if_else(is.na(pop), 0.5 ,pop)) %>%
        filter( date <= (as.Date('2006-12-01' ) %m+% months(N_advance))) %>%
        arrange(VARNAME_2, date) %>%
        group_by(VARNAME_2) %>%
        mutate( m_DHF_cases=if_else(is.na(m_DHF_cases),0, m_DHF_cases), ##ALERT...assuming nov, dec missing=0
          index= row_number() , 
               m_DHF_cases_fit = if_else(index==max(index,na.rm=T), NA_real_,m_DHF_cases)) %>%
        dplyr::select(month, year, VARNAME_2, m_DHF_cases,m_DHF_cases_fit, pop) %>%
        arrange(month, year, VARNAME_2)
}

```

## Base Model
For this model, n_advance = model # - 1 because need to start at 0 for month 12, 2006
```{r}
###############################################################

#Model Fitting

###############################################################

# default of burnin = 10000, n.sample = 110000, thin = 10

mod.func <- function(N_advance) {
  ds <- advance_missing_func(N_advance) 
 
  x_model = model.matrix(~ as.factor(as.numeric(month)), data = ds)

  offset_model = log(ds$pop)


  results<-ST.CARar(m_DHF_cases_fit ~ -1 + x_model + offset(offset_model) , 
                  data = ds,

                  family = "poisson",

                  W = neighb.mat_2,

                  burnin = 10000, 

                  n.sample = 110000,

                  thin = 10,

                  prior.mean.beta = rep(0,

                                        times = ncol(x_model)),
                                        
                  prior.var.beta = rep((100^2),

                                       times = ncol(x_model)),

                  prior.tau2 = c(0.01, 0.01),
                  
                  AR = 1,

                  MALA = TRUE,

                  verbose = TRUE)
}

```

## To get fitted cases
```{r}
mod.func <- function(N_advance) {
  ds <- advance_missing_func(N_advance) 
  
  x_model = model.matrix(~ as.factor(as.numeric(month)), data = ds)
  
  offset_model = log(ds$pop) 
  
  
  results<-ST.CARar(m_DHF_cases_fit ~ -1 + x_model + offset(offset_model) , 
                    data = ds,
                    
                    family = "poisson",
                    
                    W = neighb.mat_2,
                    
                    burnin = 100, 
                    
                    n.sample = 1100,
                    
                    thin = 10,
                    
                    prior.mean.beta = rep(0.00,
                                          
                                          times = ncol(x_model)),
                    
                    prior.var.beta = rep((100.00^2),
                                         
                                         times = ncol(x_model)),
                    
                    prior.tau2 = c(0.01, 0.01),
                    
                    AR = 1,
                    
                    MALA = FALSE,
                    
                    verbose = TRUE)
  
  # save fitted values
  post_fit <- results$fitted.values
  
  # save to a file
  saveRDS(post_fit, paste0("post_fit_values_", N_advance,".rds"))
  
  return(0)
}

```

## To get 95% CI of model fits 
```{r}


samples_to_quant <- function(X){
  quant1 <- apply(X, 2, quantile, probs=c(0.025, 0.975,seq(0.01, 0.99, by=0.01)))
  return(quant1)
} 

mod.func <- function(N_advance) {
  ds <- advance_missing_func(N_advance) 
  
  x_model = model.matrix(~ as.factor(as.numeric(month)), data = ds)
  
  offset_model = log(ds$pop) 
  
  
  results<-ST.CARar(m_DHF_cases_fit ~ -1 + x_model + offset(offset_model) , 
                    data = ds,
                    
                    family = "poisson",
                    
                    W = neighb.mat_2,
                    
                    burnin = 1000, 
                    
                    n.sample = 11000,
                    
                    thin = 10,
                    
                    prior.mean.beta = rep(0.00,
                                          
                                          times = ncol(x_model)),
                    
                    prior.var.beta = rep((100.00^2),
                                         
                                         times = ncol(x_model)),
                    
                    prior.tau2 = c(0.01, 0.01),
                    
                    AR = 1,
                    
                    MALA = FALSE,
                    
                    verbose = TRUE)

  
  # save 1% of fits + 95% CI
  post_quantiles <- lapply(X = results$samples,FUN = samples_to_quant)
  
  # save only fitted values
  post_fitted_quant <- post_quantiles$fitted
  
  # save to a file
  saveRDS(post_fitted_quant, paste0("posterior_quant_", N_advance,".rds"))
  
  return(0)
}


```
